[Update] Major game update.
Project upgrade to UE 5.4.

New features:
1. Separate character action and weapon logic to plugin.
2. New weapon system with lua script support.

Improvement:
1. Integrate Motion Warping plugin for character animation.

+++ Source/DogFightGame/Private/FunctionLibrary/CommonGameFlowFunctionLibrary.cpp
@@ -167,7 +167,7 @@ ATopDownStylePlayerController* UCommonGameFlowFunctionLibrary::GetLocalPlayerCon
 {
 	if (auto World = ULuaIntegrationFunctionLibrary::GetCurrentWorld(WorldContextObject))
 	{
-		UE_LOG(LogTemp, Log, TEXT("Get Player Controller with id: %d"), GPlayInEditorID);
+		UE_LOG(LogTemp, Log, TEXT("Get Player Controller with id: %d"), (int32)GPlayInEditorID);
 		return Cast<ATopDownStylePlayerController>(GEngine->GetFirstLocalPlayerController(World));
 	}
